download_link endpoint implementation complete

+++ Backend/youtube2mp3/converter/corelib.py
@@ -1,10 +1,15 @@
 from pytube import YouTube,extract
+from pytube.exceptions import RegexMatchError
 import ffmpeg
 import os
+from pathlib import Path
 
-from elasticsearch import Elasticsearch
+from converter.coreconfig import CONFIG
 
-# TODO : additional implementation when exception case (not found youtube video, fail to convert mp4 -> mp3) needed.
+from elasticsearch import Elasticsearch, NotFoundError
+
+
+# TODO : additional implementation when exception case (not found youtube video, fail to convert mp4 -> mp3 : duplicate youtube file, race condition...) needed.
 
 class YoutubeConverter():
 
@@ -12,27 +17,36 @@ class YoutubeConverter():
 		self.download_path = download_path
 
 	def convert_youtube(self,video_url):
-		yt = YouTube(video_url)
-		title = yt.title
-		vid = extract.video_id(video_url)
+		try:
+			yt = YouTube(video_url)
+			title = yt.title
+			vid = extract.video_id(video_url)
+		except RegexMatchError:
+			return {}
 
-		yt.streams.filter(file_extension='mp4').first().download(self.download_path,title)
+		# TODO : addtional download url factory class is need. 
+		download_url = "{}://{}/{}/{}".format(CONFIG['scheme'],CONFIG['host'],"download",vid)
+		result = {"title" : title, "video_id" : vid, "download_url" : download_url}
 
-		input_filename = "{}.{}".format(title,"mp4")
-		output_filename = "{}.{}".format(title,"mp3")
+		input_filename = "{}.{}".format(vid,"mp4")
+		output_filename = "{}.{}".format(vid,"mp3")
 
 		full_input_path = os.path.join(self.download_path,input_filename)
 		full_output_path = os.path.join(self.download_path,output_filename)
 
-		stream = ffmpeg.input(full_input_path)
-		stream = ffmpeg.output(stream, full_output_path)
-		ffmpeg.run(stream)
+		input_path_handle = Path(full_input_path)
+		output_path_handle = Path(full_output_path)
 
-		os.remove(full_input_path)
+		if input_path_handle.is_file() == False:
+			yt.streams.filter(file_extension='mp4').first().download(self.download_path,vid)
+		
+		if output_path_handle.is_file() == False:
+			stream = ffmpeg.input(full_input_path)
+			stream = ffmpeg.output(stream, full_output_path)
+			ffmpeg.run(stream)
 
-		download_url = "http://{}/{}/{}".format("localhost:8000","download",vid)
-
-		result = {"title" : title, "video_id" : vid, "download_url" : download_url}
+		if input_path_handle.is_file():
+			os.remove(full_input_path)
 
 		return result
 
@@ -48,8 +62,12 @@ class DownloadAudioInfoDTO():
 		return result
 
 	def search(self,key):
-		result = self.es.get(index=self.index, doc_type=self.doc_type, id=key)['_source']
-		return result
+		result = self.es.get(index=self.index, doc_type=self.doc_type, id=key)
+		if '_source' in result:
+			return result['_source']
+		else:
+			result = {}
+			return result
